requirements.txt
paramiko
py-getch
